+++ lib/views/booking_form.dart
@@ -1,4 +1,4 @@
-import '../database/database_helper.dart';
+import '../db/database_helper.dart';
 import '../models/user.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart'; // Import the intl package for date formatting
