+++ lib/views/package_view.dart
@@ -1,5 +1,5 @@
 import 'package:assignment1/views/booking_form.dart';
-import '../database/database_helper.dart';
+import '../db/database_helper.dart';
 import '../models/homestay.dart';
 import 'login.dart';
 import 'package:flutter/material.dart';
